Changed temperature/humidity sensors to return float value instead of int.

+++ hardware/include/IHumiditySensor.h
@@ -14,7 +14,7 @@ public:
 
   virtual const std::string& getName() const = 0;
 
-  virtual int getHumidityInPercent() const = 0;
+  virtual float getHumidityInPercent() const = 0;
 };
 
 } /* namespace hardware */
